integrations/hackathons/AI-ppt-generator-agent/requirements.txt
uagents
python-pptx
requests